Refactor API routes properly
