+++ server/routes/index.js
@@ -2,27 +2,22 @@ const express = require('express');
 const router = express.Router();
 const package = require('../../package.json');
 
-var config = require('config')
+const config = require('config')
 
-// GET /api
-router.get('/', function(req, res, next) {
-  res.send({
-    name : package.name,
-    version : package.version
-  });
-});
-
-router.get('/v1', function(req, res, next) {
-  res.send({
-    apiVersion : 1.0
-  });
-});
-
+// parse config
 const social = config.get('social')
 const defaultNetwork = config.get('currencies').btc.defaultNetwork
 const isTestnet = defaultNetwork === "testnet"
 
-router.get('/v1/config', function(req, res, next) {
+// GET /api
+const version = (req, res, next) =>
+  res.send({
+    apiVersion : 1.0,
+    name : package.name,
+    version : package.version
+  })
+
+const configInfo = (req, res, next) =>
   res.send({
     apiVersion : 1.0,
     version : package.version,
@@ -30,13 +25,14 @@ router.get('/v1/config', function(req, res, next) {
     isTestnet,
     defaultNetwork,
     ...config.get('app')
-  });
-});
+  })
 
-
-router.get('/*', function(req, res) {
+const catch404 = (req, res) =>
   res.status(404)        // HTTP status 404: NotFound
       .send('Not found');
-});
 
-module.exports = router;
+module.exports = {
+  configInfo,
+  version,
+  catch404
+}
